lora_testy/requirements.txt
future==0.18.2
iso8601==1.0.2
pkg-resources==0.0.0
pyserial==3.5
PyYAML==6.0
